+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/OpenIdConstants.kt
@@ -51,8 +51,6 @@ object OpenIdConstants {
 
     const val PROOF_JWT_TYPE = "openid4vci-proof+jwt"
 
-    const val PROOF_CWT_TYPE = "openid4vci-proof+cwt"
-
     const val AUTH_METHOD_ATTEST_JWT_CLIENT_AUTH = "attest_jwt_client_auth"
 
     const val PARAMETER_PROMPT = "prompt"
@@ -76,20 +74,12 @@ object OpenIdConstants {
 
         companion object {
             private const val STRING_JWT = "jwt"
-            private const val STRING_CWT = "cwt"
         }
 
         /**
          * Proof type in [at.asitplus.openid.CredentialRequestProof]
          */
         object JWT : ProofType(STRING_JWT)
-
-        /**
-         * Proof type in [at.asitplus.openid.CredentialRequestProof]
-         *
-         * Removed in OID4VCI Draft 14, kept here for a bit of backwards-compatibility
-         */
-        object CWT : ProofType(STRING_CWT)
 
         /**
          * Any proof type not natively supported by this library
@@ -102,7 +92,6 @@ object OpenIdConstants {
 
             override fun deserialize(decoder: Decoder): ProofType = when (val str = decoder.decodeString()) {
                 STRING_JWT -> JWT
-                STRING_CWT -> CWT
                 else -> Other(str)
             }